# Description
Enable prebuffer in the pyarrow.dataset.ParquetFragmentScanOptions. Relevant PR in [Arrow ](https://github.com/apache/arrow/pull/37854)repo, where they changed it to be default behavior. However, this won't be the case for older versions for PyArrow, so we need to set it to True.: 


It improves read speed by 6-7x on Azure in one dataset that I have.

Before:
1min 4s ± 3.48 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

After:
8.99 s ± 786 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)


# Related Issue(s)
<!---
For example:

- closes #106
--->

Closes #https://github.com/delta-io/delta-rs/issues/1569